keep doing guessNumber
