+++ rn-first-app/components/Card.js
